requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
clickhouse-connect==0.8.18
cryptography==45.0.4
fastapi==0.115.14
h11==0.16.0
idna==3.10
lz4==4.4.4
numpy==2.3.1
numpydantic==1.6.9
pandas==2.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
ruff==0.12.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uttlv==0.7.1
uvicorn==0.34.3
zstandard==0.23.0
